Updated all routers and controllers using OM. Standardized TODO formats.

+++ restful-api/controllers/watchController.js
@@ -1,5 +1,5 @@
 const Redis = require("../init/initRedis");
-const {EntityId} = require("redis-om");
+const { EntityId } = require("redis-om");
 const TODO = require("../init/initTODO");
 
 exports.getUserTasks = async (req, res) => {
@@ -8,19 +8,6 @@ exports.getUserTasks = async (req, res) => {
 
     const user = req.params.user;
     console.log(user);
-
-    // const exists = await redis.exists(user);
-
-    // if (exists) {
-
-    //     const tasks = await redis.get(user);
-    //     res.status(200).send(tasks);
-
-    // } else {
-
-    //     return res.status(404).send("The user does not exist!");
-
-    // }
 
     const tasks = await todoRepo.search().where('assignee').equals(user).return.all();
 
@@ -35,49 +22,42 @@ exports.getUserTasks = async (req, res) => {
     };
 }
 
+exports.completeUserTasks = async (req, res) => { 
+    await Redis.getConnection();
+    const todoRepo = Redis.todoRepository;
+
+    const data = req.body;
+
+    const todo = await todoRepo.search().where('title').equals(data.title).and('assignee').equals(data.assignee).return.first();
+    if (todo !== null) {
+        todo.isDone = true;
+
+        todoRepo.save(todo);
+
+        res.status(200).send("Task completed successfully.");
+    } else {
+        res.status(404).send("Task not found.");
+    }
+    
+}
+
 exports.deleteUserTasks = async (req, res) => {
 
     await Redis.getConnection();
     const todoRepo = Redis.todoRepository;
     
-    const user = req.params.user;
-    console.log(user);
-    //const tasks = req.body.tasks;
+    const data = req.body;
 
-    const todos = await todoRepo.search().where('assignee').equals(user).return.all();
-    for (let todo of todos) {
+    const todo = await todoRepo.search().where('title').equals(data.title).and('assignee').equals(data.assignee).return.first();
+
+    if (todo !== null) {
         const id = todo[EntityId];
+
         await todoRepo.remove(id);
+
+        res.status(200).send("Task deleted successfully.");
+        
+    } else {
+        res.status(404).send("Task not found.");
     }
-
-    res.status(200).send("Tasks deleted successfully.");
-    // const reply = await redis.exists(user);
-
-    // if (reply != 1) {
-
-    //     return res.status(404).send("The user does not exist!");
-
-    // } else {
-
-    //     for (task of tasks) {
-
-    //         const exists = await redis.sIsMember(user, task);
-
-    //         if (exists) {
-
-    //             await redis.sRem(user, task);
-
-    //             console.log(`${task} completed for ${user}.`);
-
-    //             await redis.sAdd("completed", task);
-
-    //         } else {
-
-    //             console.log(`${task} does not exist in ${user}'s TODO list.`);
-
-    //         }
-    //     }
-
-    //     res.status(200).send("Your tasks are registered to be completed.");
-    // }
 }
